Moved network functionality.
All the network functionality was in the main tpcengine package making the
package pretty crowded. All the network functionality has been moved to a
'net' sub package.
