+++ hazelcast-tpc-engine/src/main/java/com/hazelcast/internal/tpcengine/net/AsyncServerSocketMetrics.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.tpcengine;
+package com.hazelcast.internal.tpcengine.net;
 
 import com.hazelcast.internal.tpcengine.util.UnsafeLocator;
 import sun.misc.Unsafe;
