+++ hazelcast-tpc-engine/src/main/java/com/hazelcast/internal/tpcengine/net/AsyncSocket.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.tpcengine;
+package com.hazelcast.internal.tpcengine.net;
 
+import com.hazelcast.internal.tpcengine.Reactor;
 import com.hazelcast.internal.tpcengine.iobuffer.IOBuffer;
 
 import java.io.IOException;
