+++ hazelcast-tpc-engine/src/main/java/com/hazelcast/internal/tpcengine/net/AsyncSocketOptions.java
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.tpcengine;
+package com.hazelcast.internal.tpcengine.net;
+
+import com.hazelcast.internal.tpcengine.Option;
 
 /**
  * Options for the {@link AsyncSocket} and {@link AsyncServerSocket}.
