+++ hazelcast-tpc-engine/src/main/java/com/hazelcast/internal/tpcengine/net/ReadHandler.java
@@ -14,8 +14,11 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.tpcengine;
+package com.hazelcast.internal.tpcengine.net;
 
+
+import com.hazelcast.internal.tpcengine.Eventloop;
+import com.hazelcast.internal.tpcengine.Reactor;
 
 import java.nio.ByteBuffer;
