+++ hazelcast-tpc-engine/src/main/java/com/hazelcast/internal/tpcengine/nio/NioAsyncServerSocketBuilder.java
@@ -16,9 +16,9 @@
 
 package com.hazelcast.internal.tpcengine.nio;
 
-import com.hazelcast.internal.tpcengine.AcceptRequest;
-import com.hazelcast.internal.tpcengine.AsyncServerSocket;
-import com.hazelcast.internal.tpcengine.AsyncServerSocketBuilder;
+import com.hazelcast.internal.tpcengine.net.AcceptRequest;
+import com.hazelcast.internal.tpcengine.net.AsyncServerSocket;
+import com.hazelcast.internal.tpcengine.net.AsyncServerSocketBuilder;
 import com.hazelcast.internal.tpcengine.Option;
 
 import java.io.IOException;
