+++ hazelcast-tpc-engine/src/test/java/com/hazelcast/internal/tpcengine/ReactorTest.java
@@ -16,6 +16,9 @@
 
 package com.hazelcast.internal.tpcengine;
 
+import com.hazelcast.internal.tpcengine.net.AsyncServerSocket;
+import com.hazelcast.internal.tpcengine.net.AsyncSocket;
+import com.hazelcast.internal.tpcengine.net.ReadHandler;
 import org.junit.After;
 import org.junit.Test;
