+++ hazelcast-tpc-engine/src/test/java/com/hazelcast/internal/tpcengine/net/SocketTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.tpcengine;
+package com.hazelcast.internal.tpcengine.net;
 
 import org.junit.Test;
